requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-multipart
requests
python-dotenv
